Feat sc-8646: Delete BlockchainApi class and instead implement its changes in BCoinApi.
